+++ CHANGES.md
@@ -63,6 +63,7 @@
 -   Fixed bug where dataset ids where "undefined" in distribution URLs.
 -   Corrected incorrect source-link-status aspect name in UI dataset request URL
 -   Updated URL of City of Launceston connector.
+-   Keep search text in synch.
 
 ## 0.0.43